PP-10485 Separate out resource tests for searching agreements
Make the resource tests a bit more manageable as I'm going to be adding
more by separating out the tests for searching agreements.
